1.discussion answer,create,update,delete (#70)
2. reply to discussion answer
3.discussion answer vote

+++ src/modules/company-discussion/discussion-answer/services/discussion-answer.service.ts
@@ -3,6 +3,7 @@ import { customError } from 'src/common/errors';
 import { COMPANY_DISCUSSION_CODE } from 'src/common/errors/error.code';
 import { COMPANY_DISCUSSION_MESSAGE } from 'src/common/errors/error.message';
 import { STATUS_CODE } from 'src/common/errors/error.statusCode';
+import { FollowCompanyService } from 'src/modules/follow-unfollow-company/services/follow-company.service';
 import ConnectionArgs from 'src/modules/prisma/resolvers/pagination/connection.args';
 import { CompanyDiscussionRepository } from '../../discussion/repository/company-discussion.repository';
 import { DiscussionAnswerVoteInput } from '../dto/discussion-answer-vote';
@@ -16,6 +17,7 @@ import { DiscussionAnswerVotePayload } from '../entities/discussion-answer-vote.
 import {
   DiscussionAnswerDeletePayload,
   DiscussionAnswerPayload,
+  DiscussionAnswerReplyPayload,
 } from '../entities/discussion-answer.payload';
 import { DiscussionAnswerRepository } from '../repository/discussion-answer.repository';
 
@@ -24,6 +26,8 @@ export class DiscussionAnswerService {
   constructor(
     private readonly discussionAnswerRepository: DiscussionAnswerRepository,
     private readonly discussionRepository: CompanyDiscussionRepository,
+    private readonly companyDiscussionRepository: CompanyDiscussionRepository,
+    private readonly followCompanyService: FollowCompanyService,
   ) {}
   async getDiscussionAnswer(
     discussionId: string,
@@ -51,6 +55,23 @@ export class DiscussionAnswerService {
           COMPANY_DISCUSSION_CODE.NOT_FOUND,
           STATUS_CODE.NOT_FOUND,
         );
+      const checkOwner = await this.companyDiscussionRepository.checkOwner(
+        answer.discussionId,
+        userId,
+      );
+      if (checkOwner)
+        return this.discussionAnswerRepository.createAnswer(answer, userId);
+      const followedCompany =
+        await this.followCompanyService.checkIfUserFollowCompany(
+          discussion.companyId,
+          userId,
+        );
+      if (!followedCompany)
+        return customError(
+          COMPANY_DISCUSSION_MESSAGE.COMPANY_NOT_FOLLOWED,
+          COMPANY_DISCUSSION_CODE.COMPANY_NOT_FOLLOWED,
+          STATUS_CODE.BAD_CONFLICT,
+        );
       return this.discussionAnswerRepository.createAnswer(answer, userId);
     } catch (err) {
       throw new Error(err);
@@ -63,6 +84,15 @@ export class DiscussionAnswerService {
     userId: string,
   ): Promise<DiscussionAnswerPayload> {
     try {
+      const checkOwner = await this.discussionAnswerRepository.checkOwner(
+        id,
+        userId,
+      );
+      if (checkOwner)
+        return await this.discussionAnswerRepository.updateAnswer(
+          updateAnswer,
+          id,
+        );
       const discussionAnswer =
         await this.discussionAnswerRepository.getDiscussionAnswerByIdAndUserId(
           id,
@@ -73,6 +103,17 @@ export class DiscussionAnswerService {
           COMPANY_DISCUSSION_MESSAGE.DISCUSSION_ANSWER_NOT_FOUND,
           COMPANY_DISCUSSION_CODE.DISCUSSION_ANSWER_NOT_FOUND,
           STATUS_CODE.NOT_FOUND,
+        );
+      const followedCompany =
+        await this.followCompanyService.checkIfUserFollowCompany(
+          discussionAnswer.discussion.companyId,
+          userId,
+        );
+      if (!followedCompany)
+        return customError(
+          COMPANY_DISCUSSION_MESSAGE.COMPANY_NOT_FOLLOWED,
+          COMPANY_DISCUSSION_CODE.COMPANY_NOT_FOLLOWED,
+          STATUS_CODE.BAD_CONFLICT,
         );
       return await this.discussionAnswerRepository.updateAnswer(
         updateAnswer,
@@ -88,6 +129,12 @@ export class DiscussionAnswerService {
     userId: string,
   ): Promise<DiscussionAnswerDeletePayload> {
     try {
+      const checkOwner = await this.discussionAnswerRepository.checkOwner(
+        id,
+        userId,
+      );
+      if (checkOwner)
+        return await this.discussionAnswerRepository.deleteAnswer(id);
       const discussionAnswer =
         await this.discussionAnswerRepository.getDiscussionAnswerByIdAndUserId(
           id,
@@ -99,13 +146,54 @@ export class DiscussionAnswerService {
           COMPANY_DISCUSSION_CODE.DISCUSSION_ANSWER_NOT_FOUND,
           STATUS_CODE.NOT_FOUND,
         );
+      const followedCompany =
+        await this.followCompanyService.checkIfUserFollowCompany(
+          discussionAnswer.discussion.companyId,
+          userId,
+        );
+      if (!followedCompany)
+        return customError(
+          COMPANY_DISCUSSION_MESSAGE.COMPANY_NOT_FOLLOWED,
+          COMPANY_DISCUSSION_CODE.COMPANY_NOT_FOLLOWED,
+          STATUS_CODE.BAD_CONFLICT,
+        );
       return await this.discussionAnswerRepository.deleteAnswer(id);
     } catch (err) {
       throw new Error(err);
     }
   }
 
-  async replyToAnswer(input: ReplyToAnswerInput, userId: string) {
+  async replyToAnswer(
+    input: ReplyToAnswerInput,
+    userId: string,
+  ): Promise<DiscussionAnswerReplyPayload> {
+    const answer =
+      await this.discussionAnswerRepository.getDiscussionAnswerById(
+        input.repliedToAnswerId,
+      );
+    if (!answer)
+      return customError(
+        COMPANY_DISCUSSION_MESSAGE.DISCUSSION_ANSWER_NOT_FOUND,
+        COMPANY_DISCUSSION_CODE.DISCUSSION_ANSWER_NOT_FOUND,
+        STATUS_CODE.NOT_FOUND,
+      );
+    const checkOwner = await this.discussionAnswerRepository.checkOwner(
+      input.repliedToAnswerId,
+      userId,
+    );
+    if (checkOwner)
+      return await this.discussionAnswerRepository.replyToAnswer(input, userId);
+    const followedCompany =
+      await this.followCompanyService.checkIfUserFollowCompany(
+        answer.discussion.companyId,
+        userId,
+      );
+    if (!followedCompany)
+      return customError(
+        COMPANY_DISCUSSION_MESSAGE.COMPANY_NOT_FOLLOWED,
+        COMPANY_DISCUSSION_CODE.COMPANY_NOT_FOLLOWED,
+        STATUS_CODE.BAD_CONFLICT,
+      );
     return await this.discussionAnswerRepository.replyToAnswer(input, userId);
   }
 
@@ -114,17 +202,76 @@ export class DiscussionAnswerService {
     userId: string,
   ): Promise<DiscussionAnswerVotePayload> {
     try {
-      const discussion =
+      const checkOwner = await this.discussionAnswerRepository.checkOwner(
+        input.discussionAnswerId,
+        userId,
+      );
+      if (checkOwner)
+        return this.discussionAnswerRepository.createAnswerVote(input, userId);
+
+      const discussionAnswer =
         await this.discussionAnswerRepository.getDiscussionAnswerById(
           input.discussionId,
         );
-      if (!discussion)
+      if (!discussionAnswer)
         return customError(
           COMPANY_DISCUSSION_MESSAGE.DISCUSSION_ANSWER_NOT_FOUND,
           COMPANY_DISCUSSION_CODE.DISCUSSION_ANSWER_NOT_FOUND,
           STATUS_CODE.NOT_FOUND,
         );
+      const followedCompany =
+        await this.followCompanyService.checkIfUserFollowCompany(
+          discussionAnswer.discussion.companyId,
+          userId,
+        );
+      if (!followedCompany)
+        return customError(
+          COMPANY_DISCUSSION_MESSAGE.COMPANY_NOT_FOLLOWED,
+          COMPANY_DISCUSSION_CODE.COMPANY_NOT_FOLLOWED,
+          STATUS_CODE.BAD_CONFLICT,
+        );
       return this.discussionAnswerRepository.createAnswerVote(input, userId);
+    } catch (err) {
+      throw new Error(err);
+    }
+  }
+  async createAnswerDownvote(
+    input: DiscussionAnswerVoteInput,
+    userId: string,
+  ): Promise<DiscussionAnswerVotePayload> {
+    try {
+      const checkOwner = await this.discussionAnswerRepository.checkOwner(
+        input.discussionAnswerId,
+        userId,
+      );
+      if (checkOwner)
+        return this.discussionAnswerRepository.createAnswerVote(input, userId);
+
+      const discussionAnswer =
+        await this.discussionAnswerRepository.getDiscussionAnswerById(
+          input.discussionId,
+        );
+      if (!discussionAnswer)
+        return customError(
+          COMPANY_DISCUSSION_MESSAGE.DISCUSSION_ANSWER_NOT_FOUND,
+          COMPANY_DISCUSSION_CODE.DISCUSSION_ANSWER_NOT_FOUND,
+          STATUS_CODE.NOT_FOUND,
+        );
+      const followedCompany =
+        await this.followCompanyService.checkIfUserFollowCompany(
+          discussionAnswer.discussion.companyId,
+          userId,
+        );
+      if (!followedCompany)
+        return customError(
+          COMPANY_DISCUSSION_MESSAGE.COMPANY_NOT_FOLLOWED,
+          COMPANY_DISCUSSION_CODE.COMPANY_NOT_FOLLOWED,
+          STATUS_CODE.BAD_CONFLICT,
+        );
+      return this.discussionAnswerRepository.createAnswerDownvote(
+        input,
+        userId,
+      );
     } catch (err) {
       throw new Error(err);
     }
